requirements.txt
flask
flask-cors
numpy
pandas
scikit-learn
matplotlib
requests
python-dotenv
scipy